Move folders and files to the new lib folder.
