+++ src/lib/data-access/json-file-database.ts
@@ -1,12 +1,12 @@
 import { Constructor } from "../common/types/constructor-type";
-import { ArmorItem } from "../features/gear/armor-item";
-import { EquipmentItem } from "../features/gear/equipment-item";
+import { ArmorItem } from "../../features/gear/armor-item";
+import { EquipmentItem } from "../../features/gear/equipment-item";
 import { IDatabase } from "../common/data-access/database-interface";
 import { DbSet } from "../common/data-access/db-set";
-import { WeaponItem } from "../features/gear/weapon-item";
-import armorData from "../database/json/armor.data.json";
-import equipmentData from "../database/json/equipment.data.json";
-import weaponData from "../database/json/weapon.data.json";
+import { WeaponItem } from "../../features/gear/weapon-item";
+import armorData from "../../database/json/armor.data.json";
+import equipmentData from "../../database/json/equipment.data.json";
+import weaponData from "../../database/json/weapon.data.json";
 
 export class JsonFileDatabase implements IDatabase {
     private _dbSets: Map<string, DbSet<any>>;
