+++ src/lib/infrastructure/global.d.ts
