+++ src/lib/services/credits-abbreviator.ts
