+++ src/lib/services/path-service.ts
